bunch of refactoring
